requirements.txt
requests
beautifulsoup4
pdfplumber
psycopg2-binary
